+++ NEWS.md
@@ -1,9 +1,14 @@
-dads v0.1.0 (2020-09-28) 
+dads v0.1.0 (2020-11-13) 
 =========================
 
 ### NEW FEATURES
  
- * INIT
+ * `dads` function first release.
+ * `make.traits` function first release.
+ * `make.modifiers` function first release.
+ * `print` and `plot` functions for `"dads"` objects.
+ * `parent.traits` utility function first release.
+ * Manual first release
 
 ### MINOR IMPROVEMENTS